feat: moved theme creation to index.js file
