+++ src/index.js
@@ -1,16 +1,38 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
+import {
+  CssBaseline,
+  MuiThemeProvider,
+  createMuiTheme,
+} from '@material-ui/core';
+import { blueGrey } from '@material-ui/core/colors';
 
 import './index.css';
 import App from './App';
 
 import store from 'store';
 
+const THEME = createMuiTheme({
+  palette: {
+    background: {
+      default: blueGrey[50],
+    },
+  },
+  typography: {
+    fontFamily: ['Roboto', '"Helvetica Neue"', 'Arial'].join(','),
+    fontSize: 12,
+  },
+});
+
+
 ReactDOM.render(
   // <React.StrictMode>
   <Provider store={store}>
-    <App />
+    <MuiThemeProvider theme={THEME}>
+      <CssBaseline />
+      <App />
+    </MuiThemeProvider>
   </Provider>,
   // </React.StrictMode>
   document.getElementById('root')
